refactor: 이벤트 리스너 메서드를 bind 메서드로 분리
